pyproject.toml
[tool.poetry]
name = "dafiti_challenge"
version = "0.1.0"
description = ""
authors = ["Victor Rafael <victorsls@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.21.3"
Pillow = "^9.2.0"
djangorestframework-simplejwt = "^5.2.0"
pandas = "^1.5.0"
python-decouple = "^3.6"
dj-database-url = "^1.0.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.20.0"
black = "^22.8.0"
autopep8 = "^1.7.0"
autoflake = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
